bg, transform, graph width and height, clipping
