+++ src/graph-handler.ts
@@ -15,8 +15,6 @@ export class GraphHandler<Node extends GraphNode, Edge extends GraphEdge> {
   private renderer: GraphRenderer<Node, Edge>;
 
   private vp: [number, number] = [0, 0];
-  private cp: [number, number] = [0, 0];
-  private dp: [number, number] = [0, 0];
 
   constructor(
     view: GraphView<Node, Edge>,
@@ -28,117 +26,48 @@ export class GraphHandler<Node extends GraphNode, Edge extends GraphEdge> {
     this.renderer = renderer;
   }
 
-  moveBy = () => {
-    this.view.moveBy(this.dp[0], this.dp[1]);
-  };
-
   handleMouseMove = (e: MouseEvent) => {
     const { moveNodeIds, moveX, moveY, dragLineSourceNode, isMovingView } =
       this.state;
 
-    // this.dp[0] = e.movementX;
-    // this.dp[1] = e.movementY;
-
     if (isMovingView && !dragLineSourceNode && moveNodeIds.length === 0) {
-      // requestAnimationFrame(() => {
       this.view.moveBy(e.movementX, e.movementY);
-      // });
 
       return;
     }
 
-    const vp = this.view.getViewPosFromWindowPos(e.x, e.y);
+    this.view.viewPosFromWindowPos(this.vp, e.x, e.y);
 
-    if (dragLineSourceNode) {
-      this.state.dragLineX = vp[0];
-      this.state.dragLineY = vp[1];
+    // if (dragLineSourceNode) {
+    //   this.state.dragLineX = vp[0];
+    //   this.state.dragLineY = vp[1];
 
-      requestAnimationFrame(this.renderer.drawDragLine);
-    }
+    //   requestAnimationFrame(this.renderer.drawDragLine);
+    // }
 
-    if (moveNodeIds.length === 0) {
-      requestAnimationFrame(() => this.checkHover(vp[0], vp[1]));
+    // if (moveNodeIds.length === 0) {
+    //   requestAnimationFrame(() => this.checkHover(vp[0], vp[1]));
 
-      return;
-    }
+    //   return;
+    // }
 
-    const dx = vp[0] - moveX;
-    const dy = vp[1] - moveY;
+    // const dx = vp[0] - moveX;
+    // const dy = vp[1] - moveY;
 
-    this.state.moveX = vp[0];
-    this.state.moveY = vp[1];
+    // this.state.moveX = vp[0];
+    // this.state.moveY = vp[1];
 
-    this.renderer.clearMove();
+    // this.renderer.clearMove();
 
-    for (const id of moveNodeIds) {
-      this.view.moveNode(id, dx, dy);
-    }
+    // for (const id of moveNodeIds) {
+    //   this.view.moveNode(id, dx, dy);
+    // }
   };
 
-  private isEdgeHovered(x: number, y: number, edge: Edge) {
-    const { edgeCtx, drawData } = this.state;
-
-    const data = drawData[edge.id] as EdgeDrawData;
-
-    return (
-      edgeCtx.isPointInPath(data.path, x, y) ||
-      edgeCtx.isPointInStroke(data.linePath, x, y) ||
-      edgeCtx.isPointInPath(data.arrowPath, x, y)
-    );
-  }
-
-  private isNodeHovered(x: number, y: number, node: Node) {
-    const { nodeCtx, drawData } = this.state;
-
-    const data = drawData[node.id] as NodeDrawData;
-
-    return nodeCtx.isPointInPath(data.path, x, y);
-  }
-
-  private checkHover(vx: number, vy: number) {
-    const { nodes, edges } = this.state;
-
-    const prevId = this.state.hoveredId;
-    this.state.hoveredId = 0;
-
-    const [cx, cy] = this.view.getCanvasPosFromViewPos(vx, vy);
-    this.state.quad.getDataInRegion(vx - 1, vy - 1, 2, 2, this.state.drawIds);
-
-    for (const id of this.state.drawIds) {
-      if (nodes[id] && this.isNodeHovered(cx, cy, nodes[id]))
-        this.state.hoveredId = id;
-
-      if (edges[id] && this.isEdgeHovered(cx, cy, edges[id]))
-        this.state.hoveredId = id;
-    }
-
-    if (this.state.hoveredId === prevId) return;
-
-    const currId = this.state.hoveredId;
-    const prev = this.state.drawData[prevId];
-    const curr = this.state.drawData[currId];
-
-    if (prev) {
-      if (prev.type === GraphDataType.NODE) {
-        this.renderer.drawNode(nodes[prevId]);
-      } else if (prev.type === GraphDataType.EDGE) {
-        this.renderer.drawEdge(edges[prevId]);
-      }
-    }
-
-    if (curr) {
-      if (curr.type === GraphDataType.NODE) {
-        this.renderer.drawNode(nodes[currId]);
-      } else if (curr.type === GraphDataType.EDGE) {
-        this.renderer.drawEdge(edges[currId]);
-      }
-    }
-  }
-
   handleWheel = (e: WheelEvent) => {
-    const pos = this.view.getViewPosFromWindowPos(e.x, e.y);
+    this.view.viewPosFromWindowPos(this.vp, e.x, e.y);
 
-    this.view.zoomBy(-e.deltaY * 0.001, pos[0], pos[1]);
+    this.view.zoomBy(-e.deltaY * 0.001, this.vp[0], this.vp[1]);
   };
 
   init() {
